+++ BookLibrary.Management.BusinessLogic/BorrowService/IBorrowingFeeCalculation.cs
@@ -4,6 +4,6 @@ namespace BookLibrary.Management.BusinessLogicLayer.BorrowService
 {
     public interface IBorrowingFeeCalculation
     {
-        double GetFee(DateTime startDate, DateTime endDate);
+        decimal GetFee(DateTime startDate, DateTime endDate);
     }
 }
